refactored the config file logic.
It is now using a YAML file (but the legacy json format is still supported and read if no yaml is provided).
see etc/LocalConfigTemplate.yml and etc/GridConfigTemplate.yml for examples.
only one type of runner can be specified at a time. We could probably have a list of runners, but then we need to modify the behavior of DriverProvider.

This implementation also supports command line overrides to specify the config to use:
mvn -DrmtestConfig=/path/to/your/LocalConfigxxx.yml test

or just override a property:
mvn -Drunner.usePhantomJS=false test

the overridable properties are picked up by reflection from the Configuration object. Currently it is:

common:
  autoCloseDrivers
  jsonReportSavePath
  rmReportIP
  rmReportLivePort
  runner

local:
  runner.android.home
  runner.android.toolsVersion
  runner.useChrome
  runner.useFirefox
  runner.usePhantomJS

grid:
  runner.enableLiveStream
  runner.hubIp
  runner.localIp

while overriding an object, the quoted yaml (or json) can be given, for example:

-Drunner='type: local\nuseChrome: true'